safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowmediafrom-expected.txt

PASS media is allowed when no policy exists.
PASS media is not allowed when policy exists.
